habitat: Fix building with hab
Setting the `scaffolding_go_gopath` variable to `/src/` was a
workaround for some bug in the go scaffolding. The bug is fixed
upstream in the way that setting this variable to some arbitrary value
like it is done now actually breaks the build. So just remove the
assignment - `hab pkg build` works just fine without it.